Texture problem solved

+++ bonus/includes/map.h
@@ -6,7 +6,7 @@
 /*   By: lde-merc <lde-merc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/24 14:58:46 by luclgdm           #+#    #+#             */
-/*   Updated: 2025/06/06 15:27:46 by lde-merc         ###   ########.fr       */
+/*   Updated: 2025/06/16 14:27:07 by lde-merc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,11 +24,11 @@
 
 typedef struct s_wall
 {
-	float	height;
-	float	start;
+	double	height;
+	double	start;
 	int		column_x;
-	float	ty_step;
-	float	ty_off;
+	double	ty_step;
+	double	ty_off;
 }	t_wall;
 
 typedef struct s_door
@@ -43,12 +43,12 @@ typedef struct s_player
 {
 	t_position	pos;
 	char		direction;
-	float		angle;
-	float		dx;
-	float		dy;
-	float		velocity;
-	float		base_velocity;
-	float		rotation_speed;
+	double		angle;
+	double		dx;
+	double		dy;
+	double		velocity;
+	double		base_velocity;
+	double		rotation_speed;
 	bool		can_teleport;
 	bool		on_door;
 }	t_player;
@@ -70,10 +70,10 @@ typedef struct s_mini_map
 {
 	int		dx;
 	int		dy;
-	float	dist;
-	float	angle;
-	float	map_x;
-	float	map_y;
+	double	dist;
+	double	angle;
+	double	map_x;
+	double	map_y;
 	int		cell_x;
 	int		cell_y;
 } t_mini_map;
