+++ bonus/srcs/Gameplay/move_use.c
@@ -6,13 +6,13 @@
 /*   By: lde-merc <lde-merc@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/28 12:31:26 by lde-merc          #+#    #+#             */
-/*   Updated: 2025/06/16 12:35:43 by lde-merc         ###   ########.fr       */
+/*   Updated: 2025/06/16 15:56:38 by lde-merc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3d.h"
 
-int	ft_is_wall_at(t_game *game, float x, float y)
+int	ft_is_wall_at(t_game *game, double x, double y)
 {
 	int	map_x;
 	int	map_y;
@@ -26,7 +26,7 @@ int	ft_is_wall_at(t_game *game, float x, float y)
 		&& game->map->map[map_y][map_x] != 'T');
 }
 
-int	ft_can_move_to(t_game *game, float x, float y)
+int	ft_can_move_to(t_game *game, double x, double y)
 {
 	if (ft_is_wall_at(game, x + PLAYER_RADIUS, y))
 		return (0);
@@ -41,7 +41,7 @@ int	ft_can_move_to(t_game *game, float x, float y)
 
 void	ft_handle_left(t_game *game)
 {
-	game->player->angle -= 100 *game->player->rotation_speed;
+	game->player->angle -= 100 * game->player->rotation_speed;
 	if (game->player->angle < 0)
 		game->player->angle += 2 * PI;
 	game->player->dx = 5 * cos(game->player->angle);
